Fix benchmarks.
